made a proper package for git.server
